requirements.txt
torch==1.11.0
numpy==1.22.3
graphviz==0.19.1
requests==2.27.1
matplotlib==3.5.1
pygame==2.1.2
python-dotenv==0.20.0
